requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
chardet==4.0.0
discord.py==1.7.3
dnspython==2.1.0
idna==3.2
mongoengine==0.23.1
multidict==5.1.0
pymongo==3.12.0
typing-extensions==3.10.0.0
yapf==0.31.0
yarl==1.6.3
